deleted unnecessary files, clear OOP structure
