+++ jentic_agents/reasoners/__init__.py
@@ -1,18 +1,13 @@
 """Reasoners package - Various reasoning strategies for AI agents."""
 
-# Existing reasoners (backward compatibility)
+# Clean OOP hierarchy: BaseReasoner -> Concrete reasoners
 from .base_reasoner import BaseReasoner, ReasoningResult, StepType
 from .standard_reasoner import StandardReasoner
 from .bullet_list_reasoner import BulletPlanReasoner
 from .freeform_reasoner import FreeformReasoner
 from .hybrid_reasoner import HybridReasoner
 
-# New infrastructure (optional imports)
-from .core import AbstractReasoner, EscalationMixin, ToolExecutionMixin, MemoryIntegrationMixin
-from .interfaces import ReactInterface
-
 __all__ = [
-    # Original exports
     "BaseReasoner",
     "ReasoningResult", 
     "StepType",
@@ -20,10 +15,4 @@ __all__ = [
     "BulletPlanReasoner",
     "FreeformReasoner",
     "HybridReasoner",
-    # New infrastructure
-    "AbstractReasoner",
-    "EscalationMixin",
-    "ToolExecutionMixin", 
-    "MemoryIntegrationMixin",
-    "ReactInterface",
 ]
