--- jentic_agents/reasoners/interfaces/__init__.py
@@ -1,7 +1,0 @@
-"""Reasoner interfaces and contracts."""
-
-from .react_interface import ReactInterface
-
-__all__ = [
-    "ReactInterface",
-]
